requirements.txt
discord
discord.py
discord-pretty-help
colorama
aiohttp